+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
-…
+
+### Breaking Changes
+ * Added method `Type\TableSchema::base_name()`
 
 ## [0.1.1]